fix:fix edit roomtype function

+++ server/utils/upload.js
@@ -19,28 +19,33 @@ const cloudinaryUpload = async (files) => {
 
 const RoomImageUpload = async (files) => {
   const fileUrl = [];
-  for (let file of files.main_img) {
-    const result = await cloudinary.uploader.upload(file.path, {
-      folder: "techupth/Neatly-project/roomImage",
-      type: "private",
-    });
-    fileUrl.push({
-      main_url: result.secure_url,
-      publicId: result.public_id,
-    });
-    await fs.unlink(file.path);
+  if (files.main_img) {
+    for (let file of files.main_img) {
+      const result = await cloudinary.uploader.upload(file.path, {
+        folder: "techupth/Neatly-project/roomImage",
+        type: "private",
+      });
+      fileUrl.push({
+        main_url: result.secure_url,
+        publicId: result.public_id,
+      });
+      await fs.unlink(file.path);
+    }
   }
-  for (let file of files.gallery_img) {
-    const result = await cloudinary.uploader.upload(file.path, {
-      folder: "techupth/Neatly-project/roomImage",
-      type: "private",
-    });
-    fileUrl.push({
-      gal_url: result.secure_url,
-      publicId: result.public_id,
-    });
-    await fs.unlink(file.path);
+  if (files.gallery_img) {
+    for (let file of files.gallery_img) {
+      const result = await cloudinary.uploader.upload(file.path, {
+        folder: "techupth/Neatly-project/roomImage",
+        type: "private",
+      });
+      fileUrl.push({
+        gal_url: result.secure_url,
+        publicId: result.public_id,
+      });
+      await fs.unlink(file.path);
+    }
   }
+
   return fileUrl;
 };
 export { cloudinaryUpload, RoomImageUpload };
